[irods/irods_5479] Add new PAM iinit flow
When using the new PAM plugin which conforms to the authentication
plugin framework for 4.3.0, there is a new flow of which iinit needs to
be made aware. However, we need to be able to authenticate with servers
using the older plugin as well, so this change threads the needle to add
support for the new plugin while maintaining support for the old.