[irods/irods 6382] Remove password arg from iinit
The argument-supplied password for iinit is a way of "bypassing" the
password prompt. This has been removed as it is insecure and adds
unneeded complications to the authentication plugins. The password
prompt can be dealt with in a scripting situation using stdin
redirection.

This also removes the echo option from iinit, which existed to echo
the user's password as they type it. This was only applicable for
plugins which use the password obtained from iinit, which turns out to
be only the native plugin.

Also moves around some variables for readability.